refactor: deduplicate iOS/tvOS LinkerOverrides via shared uikit-common file
The shared UIKit control-preservation code was duplicated between the ios and
tvos LinkerOverrides (flagged as duplicated new code by Sonar). Move the common
controls into a single Platforms/uikit-common/LinkerOverrides.cs (already
compiled into ios, tvos and maccatalyst) and keep the iOS-only controls
(UISlider, UIRefreshControl, UISwitch) in a separate IosLinkerOverrides class.
Verified ios, tvos and maccatalyst compile 0/0 in the dockur guest.
